Update wcmTouchDevice for touch interface
To support arbitration, we only needed to access touch pointer
from pen interfaces. Now, we need to access touch pointer from
touch interfaces to update touch switch state.

Since other tools, such as PAD, can be associated with touch
interfaces, we can not assume the first tool created on touch
interface is a touch tool. Assign it to wcmTouchDevice to avoid
looping through the tool list everytime when we need to access it.

Signed-off-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>