wcmUSB: Don't directly dereference common->fd
common->fd is intended to share the fd between different InputInfo structs,
and that is the only thing it should be used for. With server managed fds,
the fd management will all be done by the server and common->fd will never
get set, so nothing outside of wcmDevOpen / wcmDevClose should use it.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>