Attempt to not lose events to pressure recalibration
Worn out devices send a non-zero pressure even when not in contact
with the tablet. To compensate for this the driver detects if the
pressure sent by the device immediately after indicating proximity
is non-zero. It subtracts this value from any pressure value sent
later and rescales the pressure range to the full device range.
If it later on sees the pressure value fall below this initial value
it will readjust it this lower value.
The downside of this is that when the pen is pushed onto the tablet
really fast the initial pressure reading may be non-zero also the
pen isn't worn. This can lead to lost events.
This patch tries to address this:
If the first pressure reading is != 0 it is recorded. If the recorded
maximum value is >0 but a later pressure reading is higher the maximum
value is replaced. If no button press event is generated the 'normal'
way it is checked of the recorded maximum would trigger one when
minPressure decreases.
Once a 'normal' button event is generated or an 'alternative' button
event is generated and minPressure doesn't decrease any more the
recorded maximum is set to 0 which will disable the checks until the
next prox in.

Signed-off-by: Egbert Eich <eich@suse.com>
Reviewed-by: Jason Gerecke <killertofu@gmail.com>