Remove duplicate tool type detection
Tool type detection is done in two places: early in usbDispatchEvents()
by calling usbInitToolType() and later on in the same function when the
events are parsed.
usbInitToolType() is used to set (wcmUSBData*)->wcmDeviceType, the
detection that happens later when the events are parsed sets
(WacomDeviceState).device_type. These variables are matched against each
to find the right channel for a device.
If the algorithms used for both set of tool type detection diverge
undesirable effects may happen. Therefore it is advisable to determine the
tool type only once ie. in usbInitToolType() and copy the result to
(WacomDeviceState).device_type if this value is unset.

* Bring the algorithm in usbInitToolType() in sync with the reset:

  Add missing key codes to deviceTypeFromEvent()

  - Add BTN_TOOL_DOUBLETAP and BTN_TOOL_TRIPLETAP, return
    TOUCH_ID for them.

  - Return TOUCH_ID on BTN_TOUCH when protocol level is
    WCM_PROTOCOL_GENERIC and tool is a 1 finger USB touch.

* Delete any duplicate tool type setting

* Add a test for debugging if the device type has changed while in proximity.

Signed-off-by: Egbert Eich <eich@suse.com>
Reviewed-by: Jason Gerecke <killertofu@gmail.com>