Do not store a second copy of old touchstrip values in oldTilt[XY]
The touchstrip values are already stored in oldStrip[XY], and appear to
also be stored in oldTilt[XY] as a quirk of history. The value of oldTilt[XY]
itself is only read in wcmSendNonPadEvents (which is not called for the pad
device for obvious reasons ;)). At this point in time, there is absolutely
no reason to store a second copy of the strip values in oldTilt[XY].

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Ping Cheng <pingc@wacom.com>