Allow touch to grab and maintain pointer control from idle puck
This patch changes the how pointer access is arbitrated. Before,
any non-touch tool would be immediately granted control over the
pointer, preempting touch itself. This is problematic if the user
happens to use a device with a puck (i.e., a Cursor device) since
the tool is likely to be in-range but idle for a significant
ammount of time.

This patch adds conditions which allow touch to grab control from
the puck if no events have been seen from it in a while. In addition,
the puck may not immediatley regain control from the touch device
while it is in use due to potential hardware jitter.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>