Add new GESTURE_CANCEL_MODE and wcmCancelGesture function
The new GESTURE_CANCEL_MODE is used to short-circuit the gesture
recognition logic until all touches have been removed from the device.
By using the wcmCancelGesture function, any in-progress gesture will
be canceled and cause the driver to ignore further touch input until a
new gesture is started.

This is used by the arbitration logic to ensure that touches which
accumulate in the valid.state list while a pen is in proximity do not
suddenly trigger a gesture at the moment the pen leaves prox.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>