Fix pointer jump when tool enters prox with filtered coordinate
When a tool enters proximity, the usbChooseChannel function zeros
out one of the WacomDeviceState structures in the common->wcmChannel
array and returns its index. That state is then initalized using
the input events recieved from the kernel, and is assumed to be
a complete description by the time it is handed off to the wcmEvent
function.

The kernel's duplicate event filtering can cause problems for this,
however. If the value of the axis as it enters prox is identical to
its value when it left prox, the kernel will not send an event for
it (leaving the associated WacomDeviceState field zero). If this
occurs with the X and/or Y axis, the pointer will jump to the wrong
screen location.

To prevent the pointer from jumping around, we check if the X or Y
value of the WacomDeviceState are still zero after reading through
an enter-prox packet. If either is, we request its current value
from the kernel just in case it was filtered out. Other axes may be
affected by this bug, but should not cause as many problems.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>