Set the active driver to NULL before uninitalizing the device
If a device being used is disconnected, it may be possible to find
yourself in a situation where a second device performs an "active"
check and dereferences a null (freed in 'wcmFree') 'priv' pointer.
To prevent this from occuring, always check if the device being
uninitialized is considered 'active' and unset it if so.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>