scan-build: security: Potential insecure memory buffer bounds restriction in call 'strcat'
The 'argv' strings passed into strjoinsplit are concatenated and
stored within the 'buff' array on the stack. Although we are
careful to ensure there is enough room remaining to store *argv
and a separating ' ' character before performing strcat, we also
need to ensure that there's room for the null terminator that
strcat itself will append.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>