Reset wcmGestureState to current device state upon gesture start
The scroll and zoom gestures use of how far the fingers have travelled
as a way to determine how many times to send the scroll/zoom event.
When a gesture first starts, however, the fingers have already moved a
large distance simply to have the gesture be recognized. This may cause
a rapid and undesired "catch up" effect due to the large built-up delta.

To work around this, when a gesture starts, we reset the wcmGestureState
so that the functions are working with a clean slate and do not see the
large delta.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>