Strengthen condition that pad may never be arbitrated pointer control
The pad should never be arbitrated pointer control since it is not a
pointer controlling device. The `check_arbitrated_control` function
performs its checks in the wrong order, however, and will grant control
if no other device is currently active.

This can cause touch to be disabled for "generic protocol" devices (e.g.
Intuos, Bamboo, etc.) until a pen is used because touch is the lowest-
priority device to be arbitrated access. For protocol 4/5 devices (e.g.
Intuos 5/4/3 and Cintiq) this can also lock out touch, but only if the
ExpressKey was pressed prior to the beginning of the touch, and only
while it remains held.

This patch strengthens the condition that the pad may never be granted
control by making it the first thing checked.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>