xsetwacom: Print proper names of xorg.conf options
Some properties have different names between xorg.conf and xsetwacom.
This adds an 'x11name' member to each parmeter which stores the name
of the associated xorg.conf property if it exists. If this member has
not been set for any particular parameter while the xorg.conf format
is selected than an error will be printed.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>