xsetwacom: Fix 'get_mapped_area' ("maptooutput next") on 64-bit
Using the "maptooutput next" command fails on 64-bit systems,
with results ranging from always mapping the pointer to the
first head to displaying errors about the transformation matrix
being non-rectangular.

This is a result of the 'get_mapped_area' function expecting
32-bit properties returned by the server to take up only 4
bytes of memory. This is the case on 32-bit systems, but not
on 64-bit ones. This was partially addressed in bc5fd9e, but
not here.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>