Modify wcmMaxCursorDist calculation to support non-inverted protocol 4
Distance values reported by the puck tool on protocol 4 tablets (e.g.
the Graphire4) have an inverted scale compared to other devices and
tools. That is to say, the values are at their maximum while the puck
is resting on the tablet and minimum just prior to going out of prox.
The way that xf86-input-wacom handles this is with explicit protocol
version checks, which limits our ability to fix this issue going
forward (when libinput/Wayland begin to rule the desktop).

This patch changes how the driver calculates wcmMaxCursorDist so that
it is able to gracefully handle the day we fix this kernel bug.
Instead of storing the minimum/maximum value ever seen on the tablet
(switching between codepaths based on protocol 5/4) we store the value
that is seen whenever the user clicks a button. It is reasonably safe
to assume that this occurs with the puck on the tablet and thus will
be the correct value. Even if the user happens to click with the puck
raised above the surface, a subsuqent click *on* the surface will
reset the value.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>