Update release.sh script from Xorg
The latest-and-greatest copy of release.sh from Xorg now replaces our
current version of the script which largely dates back to 2010. This
script brings several new features along with it, with stronger checks
and PGP signing probably being the two worth noting.

A very small number of modifications to the upstream version of this
script (commit 65cb27b) have been made to support this project. The only
one of note is a change to 'generate_announce' which changes it to
use the 'Reply-To' header instead of 'Cc' (since the announce list will
bounce unauthorized messages back).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>