xsetwacom: Remove unnecessary static state from 'get_actions'
The 'last_type' variable within 'get_actions' stores the type of the last
action encountered. When dealing with "key" or "button" actions, we use
that information to determine if we need to print out the action prefix
or not (if the type hasn't changed, its safe to leave the prefix out).

For some reason, this variable was marked as 'static', which causes it
to retain its value across invocations. The function is only called once
for any given button, meaning that we improperly retain the "last_type"
across buttons. If the last action on a button is of e.g. type "key" and
the first action of the next button is as well, then the "key" prefix
will be missing from the printed output of that second button's actions.
Making this variable non-static fixes this issue and ensures each run
of the function is independent.

Fixes: http://sourceforge.net/p/linuxwacom/bugs/303/
Fixes: https://github.com/linuxwacom/xf86-input-wacom/issues/3

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>