conf: Prevent xf86-input-wacom from binding to pure-keyboard devices
The Wacom Bluetooth Keyboard matches the "Wacom class" and so is bound to the
xf86-input-wacom driver. This driver is not designed to work with keyboards,
so we narrow the set of devices our InputClasses will match. Because we need
to match tablets, touchpads, and touchscreens (pad devices are a special case
of tablet) but there is no way to specify an "or" match we duplicate existing
classes and assign one of the above types to each.

Note that simply adding 'MatchIsKeyboard "false"' to our existing snippets
would cause devices like the Cintiq 24HDT Pad and Cintiq 27QHDT Pad from
binding to our driver since both are recognized as (partial) keyboards due
to the presence of KEY_* kernel events.

Fixes: https://sourceforge.net/p/linuxwacom/bugs/294/
Fixes: https://github.com/linuxwacom/xf86-input-wacom/issues/1

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>