conf: rename to 70-wacom.conf
This is part of a two-step solution, the other half is renaming the
xf86-input-libinput config snippet to sort lower than ours.

Currently libinput picks up devices that are (for now) destined to the wacom
driver. Since the wacom driver is more of a leaf package than libinput, the
best option here is to make the wacom driver sort higher and let users
uninstall it when not needed. To avoid crowding the 90-* space where users
usually have custom config snippets, libinput will drop down to 60 and
wacom will bump up to 70.

Fixes: https://sourceforge.net/p/linuxwacom/bugs/315/
Fixes: https://github.com/linuxwacom/xf86-input-wacom/issues/4

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>