Handle the input ABI 23 - removal of xf86BlockSIGIO
Input is now handled in a separate thread in the server when a device is added
through xf86AddEnabledDevice(). Previously that was in a sigio handler and we
have a bunch of workarounds to avoid calling things from within the signal
handler. With the new API these aren't needed, so split those callbacks and
call them directly where needed.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Ping Cheng <pingc@wacom.com>