Increase full-scale pressure range from 0..2047 to 0..65535
The driver has historically normalized the pressure range of all kernel
devices to 0..2047 rather than using their native range to keep things
like the application of the pressure curve simple. Pens that report more
than 2048 pressure levels are also normalized down to this range though,
reducing their precision. In order to accomodate the new 8K pen (and any
future pens with even higher precision), this patch bumps up the full-
scale range to be 0..65535. This number was chosen both because it far
exceeds anything currently known about, and also because it matches the
normalization range used over the wire by the Wayland tablet protocol.

Note that the WACOM_PROP_PRESSURE_THRESHOLD value has been tied to the
normalized (2048-level) pressure range for some time, meaning that we
cannot simply change the range without causing a change in the perceived
threshold for users. To ensure compatibility, the value is interpreted
as a fraction of 2048 and then scaled to the actual normalization range.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>