xsetwacom: Add return to special key alias list
return is a popular key assignment which was not added to the
special key list. Although Return would work since it is the
key name, we add return anyway to make users happy.

Enter/enter for KP_Enter could have achieved the same purpose.
But, it confuses users (and some developers ;).

Signed-off-by: Ping Cheng <pingc@wacom.com>
Tested-by: Kelly Price <strredwolf@gmail.com>