Read TouchRing axis range from kernel instead of hardcoding
The TouchRing on the MobileStudio Pro has a range of only 0-35 instead
of 0-17 like prior devices. Because we hardcode an assumed range, the
driver mistakenly believes the jump from 35 to 0 (or 0 to 35, depending
on direction) when the user completes a revolution is actually caused
by the user reversing their finger direction. By reading the range from
the kernel, we can avoid this situation.

Note that the ABS_WHEEL axis is also used by (legacy) combined pen/pad
devices with a range corresponding to the puck fingerwheel. We need to
be careful to not read the value in these cases since it would lead to
erroneous behavior in existing setups. Devices with a range range
different from 0-71 will hopefully not be widely used on kernels prior
to 3.17 where pen and pad were split into seperate devices since there
is no way (short of peeking at the VID:PID or name) to have them work
correctly in the combined case.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <pingc@wacom.com>