Lazily allocate pressure curve table
Now that the pressure curve contains 65K points it takes up quite a bit
of memory, especially considering that the pressure curve may not need
to exist for some devices (e.g. pads) and may just be the default linear
curve even for those where it should exist. To reduce the amount of
memory used, we now lazily allocate space for the pressure curve tables
only when they are set to a non-default curve.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>