Allow cursor devices to maintain "active" status while buttons are down
Cursor (puck) tools tend to be left on the sensor, preventing us from
being able to rely solely on proximity information to determine if
they are being actively used. In the past we've used the amount of time
since the last event as an indicator of activity and allowed other
devices to grab control of the pointer if more than 100 milliseconds had
elapsed since the cursor's last event. Although this seems to work well,
there is another indicator of activity which should not be ignored:
button state. If a user is pressing a button on their cursor tool, it
should be considered active even if the 100ms timeout has been exceeded.
Not doing so could potentially allow another tool to grab "active"
status and have our driver send a button-up message and stop an in-
progress drag.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>