Further reduce arbitration priority of cursor devices
Cursor (puck) devices are typically left on the sensor and can emit
spurious events that can potentially cause the driver to steal "active"
status from another tool. In the past we'd only considered the case where
the active tool was a touch but it is reasonable to extend this logic to
other tool types. If a cursor emits a spurious event in the middle of a
pen dragging, for example, the drag will momentarily stop as the pen is
temporarily sent out of prox -- with potentially disasterous results.
Additionally, when the pen "returns" in prox on the next event, any
currently-applied pressure will be taken as the preload and result in
incorrect pressure scaling until the pen is removed from contact.

This commit removes the "IsTouch" conditions from the handling of
spurious cursor events in check_arbitrated_control. The result is that a
cursor will not be granted "active" status in preference to any other
tool (and will drop "active" status if gained after 100ms and while
no buttons are pressed).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>