Introduce "Pressure2K" config option for incompatible software
It appears that some software may not be entirely compatible with the
expanded 65K pressure level range that was introduced in version 0.34.0.
Although our driver advertises the larger range in XI2 (and toolkits
like GTK+2/3 and Qt3/4/5 make use of it), there have been reports of
other software (e.g. The Foundry's "NUKE") misbehaving.

As a workaround, this patch introduces a new boolean config option named
"Pressure2K". If enabled, it causes the driver to revert to its prior
behavior of using a pressure range of 0-2047. This option is disabled by
default, but can be turned on by adding the following configuration
snippet to a new file in the /etc/X11/xorg.conf.d directory:

Section "InputClass"
    Identifier "Wacom pressure compatibility"
    MatchDriver "wacom"
    Option "Pressure2K" "true"
EndSection

Ref: https://sourceforge.net/p/linuxwacom/mailman/message/35857403/
Ref: 3e56ce4429 (Increase full-scale pressure range from 0..2047 to 0..65535)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>