test: Update test_normalize_pressure to work with maxCurve
Commit d958ab79d2 changes the driver's pressure normalization functions
to use a variable "maxCurve" which defines the upper limit of pressure
values to be emitted by the driver, rather than using FILTER_PRESSURE_RES
directly. The tests were never updated to take this into account and
fail due to maxCurve never being set.

This patch has the normalize_pressure test set the value of maxCurve.
While we're at it, we have the test verify that multiple values of
maxCurve work as expected.

Fixes: d958ab79d2 (Introduce "Pressure2K" config option for incompatible software)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>