Remove the device's fd from the select() set when we get ENODEV
If the device is unplugged, the fd triggers in select/poll/... but comes back
with ENODEV. This triggers a lot of error messages in the log until finally
the udev code catches up with us and the device is removed properly.

Catch that case by removing the fd from the select() set so we don't get
triggered to call read_input on it anymore.

https://sourceforge.net/p/linuxwacom/bugs/337/

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: crocket <crockabiscuit@gmail.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>