Add Travis-CI integration
Adds a .travis.yml file which allows the Travis-CI service to build and
test the project. This should ensure we're immediately warned of potential
issues even if someone forgets to run the test suite before pushing. The
build is performed with "-Wall -Wextra -Werror" (with only a handful of
currently-necessary exceptions) on both GCC and Clang.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>