Build tools with AM_CFLAGS and its warnings
The Makefile.am contained under the tools subdirectory makes use of an
Automake feature that allows the definition of per-executable CFLAGS.
These CFLAGS definitions do not automatically include the general-use
AM_CFLAGS. This has resulted in these executables not being built with
the intentended warnings/errors enabled. Note that we also explicitly
add "-Wno-error" to the definition of the "xsetwacom-test" binary since
we don't want warnings there to break anything.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>