Use git-version-gen for build-time version strings
Currently the driver uses macros like PACKAGE_VERSION whenever it needs
to print out a version string. These macros are only updated whenever
the autoconf.sh script is run, however. As a practical result, this means
that the version information which is built into development versions of
the driver is often inaccurate. To fix this, we define a new BUILD_VERSION
macro which is updated whenever git reports a changed version number.

In addition to having `xsetwacom -V` report the precise build, we also
have the X log print out the build number to augment the MAJOR/MINOR/PATCH
data that is normally printed.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>