Add Coverity integration to .travis.yml
Adds the ability to run a Coverity scan from Travis. Since there are a
limited number of times per week that Coverity can be used, we only
run it when Travis has the DO_COVERITY environment variable set to "YES".
This can either be set manually through the Travis' interface, but will
also be set for cron-triggered jobs.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>