Do not overwrite raw device state with filtered coordinate values
By performing coordinate filtering in wcmEvent and overwriting the raw
device state stored in pChannel, we loose the ability to see the actual
(but noisy) hardware coordinates at later points in the code. By moving
this filtering into commonDispatchDevice we can apply the result to only
the 'filtered' DeviceState sample that is ultimately passed along to
wcmSendEvents.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>