Reset filter when tip is touched to tablet
Coordinate averaging is useful for producing smooth strokes, but this
averaging can cause button events to be sent at the wrong location.
In particular, when making a series of quick strokes with an AES pen,
the actual location the pen contacts the display can be some distance
away from the first (low-quality) position report. This can result in
"hook" artifacts if it is still influencing the averaged pointer
position when the pen goes down. This can also be seen if the RawSample
parameter is increased significantly.

To prevent these artifacts from appearing, we reset the averaging filter
whenever the tip is pressed. This ensures that the event event is sent
at the pen's actual location, plus or minus some (probably smaller) noise.
In theory these artifacts could still appear if drawing were triggered
by a button other than the tip, but that would be a quite non-standard
setup.

https://sourceforge.net/p/linuxwacom/bugs/338/

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>