conf: Add preprobe fdi config to workaround EV_SW issues in HAL
The HAL daemon used by RHEL6/CentOS6 does not play nicely with input
devices that advertise unsupported EV_SW events. Devices with such
an event will fail in the middle of the probe process, preventing
the X server from being able to hotplug the device. The SW_MUTE_DEVICE
event used by the Cintiq Pro touchscreen is such an event.

To work around this limitation, we can define an fdi configuration file
which resets the "button.has_state" property prior to probe. Because
HAL only runs the failing code when this property is true, the device
is able to finish the probe process just as though it never had the
unsupported switch in the first place. Resetting this property does
prevent HAL from being able to monitor the switch state, but that is
not considered a problem.

Since HAL is long-since deprecated, it doesn't make sense to install
this configuration file on systems which don't require it. This
commit also adds logic to configure.ac so that only systems which still
use HAL will install this file.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>