Be more strict about linking to already-linked devices
The wcmLinkTouchAndPen function is responsible for associating the pen and
touch interfaces of a tablet. This association can be tripped up if multiple
devices with the same name are connected, so it includes code to prevent
linking to a device which already has been linked up. However, the condition
also checks that the device being considered is a tablet (i.e., stylus/eraser/
cursor; not pad/touch). This latter check seems nonsensical: it allows devices
to link to the first touch device they find, even if that device is already
linked up.

Disallowing multiple links to the same device prevents all three pen/touch/pad
devices from having wcmTouchDevice set up, however. Leaving the stricter
condition as-is could cause devices to be linked up incorrectly (e.g. if the
X server has {touch1, pen1, pad1, touch2, pen2, pad2}, then touch1 and pen1
would get linked, but pad1 and touch2 would get linked, and pen2/pad2 left
unlinked!). To prevent this, we disallow linking of pad interfaces, which
should not be a problem since the pad shouldn't need to use the wcmTouchDevice
variable.

Fixes: 8bb519ef2b ("Update wcmTouchDevice for touch interface")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>