Allocate 4 buttons for stylus devices (Pro Pen 3D)
The Pro Pen 3D includes a third stylus button, reported from the kernel
as BTN_STYLUS3. Support for this event was added earlier, but the number
of buttons allocated was not updated. This meant that although the X
driver would properly set the button flag bit to indicate that a button
should be sent, a userspace event wouldn't actually be sent. This commit
increases the number of buttons allocated, causing the third button to
emit X11 button number 8 when pressed. This button can also be configured
through xsetwacom now that the driver allocates the correct number.

Fixes: 4a96a58f26 ("Add support for kernel's new BTN_STYLUS3 event")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>